pyproject.toml
[project]
name = "tis-cli"
version = "0.1.0"
description = "Access the TOPMed Imputation Server programmatically."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mff-pretty-cli>=0.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.4",
    "requests-toolbelt>=1.0.0",
    "tqdm>=4.67.1",
]

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"local" = "local"}
